Move components to Home folder
